requirements.txt
streamlit==1.40.1
pandas==2.0.3
requests==2.32.4
snowflake-connector-python==3.1.6
Pillow==10.0.1
openpyxl==3.1.2
